Add Doc interface for include_docs parameter
Hi,

Thanks for this useful lib. This adds the `Doc` field so that when querying a view with the include_docs field, the docs are now accessible view the `Doc` field. 

Often a views value is not the actual doc. Instead we can add the `include_doc = true` parameter to the view query to get the docs associated with the view.